requirements.txt
asgiref==3.9.2
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
pycparser==2.23
PyJWT==2.10.1
reportlab==4.4.4
sqlparse==0.5.3
